[Refactoring]Member
회원 서비스 인터페이스 추가 및 DIP 부분 준수
