+++ OInjo_Shop/src/main/java/kr/idu/OInjo_Shop/controller/MemberController.java
@@ -9,6 +9,7 @@ import kr.idu.OInjo_Shop.service.Admin.AdminAuthenticator;
 import kr.idu.OInjo_Shop.service.Mail.MailService;
 import kr.idu.OInjo_Shop.service.Mail.RegisterMailService;
 import kr.idu.OInjo_Shop.service.Member.MemberService;
+import kr.idu.OInjo_Shop.service.Member.MemberServiceImpl;
 import kr.idu.OInjo_Shop.service.Order.OrderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -148,7 +149,7 @@ public class MemberController {
     public String updateForm(HttpSession session, Model model) {
         // 정보 수정 -> 세션에 있는 로그인 값으로 전체 정보를 DB로부터 가져와서 model에 담음
         String myEmail = (String) session.getAttribute("loginEmail"); // 캐스팅 - 강제 형변환(Object -> String)
-        MemberDTO memberDTO = memberService.updateForm(myEmail);
+        MemberDTO memberDTO = memberService.findEmail(myEmail);
         model.addAttribute("updateMember", memberDTO);
         return "member/memberupdate";
     }
